fix: 엄격한 타입 에러 해결
