+++ frontend/src/components/ChecklistCustom/CustomTabs.tsx
@@ -1,0 +1,16 @@
+import Tabs from '@/components/common/Tabs/Tabs';
+import useChecklistCustomStore from '@/store/useChecklistCustomStore';
+
+export const ChecklistCustomTabs = () => {
+  const { validCategory: validCategoryEditMode } = useChecklistCustomStore();
+
+  const checklistCustomTabs = validCategoryEditMode.map(category => ({
+    id: category.categoryId,
+    name: category.categoryName as string,
+  }));
+  return (
+    <>
+      <Tabs type="normal" tabList={checklistCustomTabs} />
+    </>
+  );
+};
