+++ frontend/src/components/common/Tabs/Tabs.stories.tsx
@@ -34,6 +34,7 @@ const mockTabsWithCompletion = newChecklistTabs.map(tab => ({ ...tab, isComplete
 
 export const Default: Story = {
   args: {
+    type: 'completion',
     tabList: mockTabsWithCompletion,
   },
 };
