+++ frontend/src/components/common/Tabs/Tabs.tsx
@@ -2,13 +2,22 @@ import styled from '@emotion/styled';
 
 import Tab from '@/components/common/Tabs/Tab';
 import { useTabContext } from '@/components/common/Tabs/TabContext';
-import { TabWithCompletion } from '@/types/tab';
 
 interface Props {
-  tabList: TabWithCompletion[];
+  tabList: TabWithCompletion[] | Tab[];
+  type: 'completion' | 'normal';
 }
 
-const Tabs = ({ tabList }: Props) => {
+export interface Tab {
+  name: string;
+  id: number;
+}
+
+export interface TabWithCompletion extends Tab {
+  isCompleted: boolean;
+}
+
+const Tabs = ({ tabList, type }: Props) => {
   const { currentTabId, setCurrentTabId } = useTabContext();
 
   const onMoveTab = (tabId: number) => {
@@ -20,7 +29,8 @@ const Tabs = ({ tabList }: Props) => {
       <Container>
         <FlexContainer>
           {tabList?.map(tab => {
-            const { isCompleted, id, name } = tab;
+            const { id, name } = tab;
+            const isCompleted = type === 'completion' && 'isCompleted' in tab ? tab.isCompleted : null;
             return (
               <Tab
                 id={id}
