+++ frontend/src/pages/NewChecklistPage/NewChecklistPage.tsx
@@ -6,11 +6,11 @@ import { getChecklistQuestions, postChecklist } from '@/apis/checklist';
 import Button from '@/components/common/Button/Button';
 import Header from '@/components/common/Header/Header';
 import { TabProvider } from '@/components/common/Tabs/TabContext';
+import NewChecklistTab from '@/components/NewChecklist/NewChecklistTab';
 import { ROUTE_PATH } from '@/constants/routePath';
 import useInputs from '@/hooks/useInputs';
 import useToast from '@/hooks/useToast';
 import NewChecklistContent from '@/pages/NewChecklistPage/NewChecklistContent';
-import NewChecklistTab from '@/pages/NewChecklistPage/NewChecklistTab';
 import useChecklistStore from '@/store/useChecklistStore';
 import useOptionStore from '@/store/useOptionStore';
 import { flexCenter, title2 } from '@/styles/common';
