--- frontend/src/types/tab.ts
@@ -1,8 +1,0 @@
-export interface Tab {
-  name: string;
-  id: number;
-}
-
-export interface TabWithCompletion extends Tab {
-  isCompleted?: boolean;
-}
